+++ additions/proxy/src/main/java/com/velocitypowered/proxy/protocol/packet/RemoveEntityEffectPacket.java
@@ -1,0 +1,70 @@
+/*
+ * Copyright (C) 2026 Velocity Contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.velocitypowered.proxy.protocol.packet;
+
+import com.velocitypowered.api.network.ProtocolVersion;
+import com.velocitypowered.proxy.connection.MinecraftSessionHandler;
+import com.velocitypowered.proxy.protocol.MinecraftPacket;
+import com.velocitypowered.proxy.protocol.ProtocolUtils;
+import io.netty.buffer.ByteBuf;
+
+/**
+ * Clientbound remove-mob-effect packet.
+ */
+public class RemoveEntityEffectPacket implements MinecraftPacket {
+
+  private int entityId;
+  private int effectId;
+
+  public RemoveEntityEffectPacket() {
+  }
+
+  public RemoveEntityEffectPacket(int entityId, int effectId) {
+    this.entityId = entityId;
+    this.effectId = effectId;
+  }
+
+  public int getEntityId() {
+    return entityId;
+  }
+
+  public void setEntityId(int entityId) {
+    this.entityId = entityId;
+  }
+
+  public int getEffectId() {
+    return effectId;
+  }
+
+  @Override
+  public void decode(ByteBuf buf, ProtocolUtils.Direction direction, ProtocolVersion version) {
+    this.entityId = ProtocolUtils.readVarInt(buf);
+    this.effectId = EntityEffectPacket.readEffectId(buf, version);
+  }
+
+  @Override
+  public void encode(ByteBuf buf, ProtocolUtils.Direction direction, ProtocolVersion version) {
+    ProtocolUtils.writeVarInt(buf, entityId);
+    EntityEffectPacket.writeEffectId(buf, version, effectId);
+  }
+
+  @Override
+  public boolean handle(MinecraftSessionHandler handler) {
+    return handler.handle(this);
+  }
+}